backend/requirements.txt
fastapi
tortoise-orm
uvicorn
